Formalise use of a factory, and remove separate router factory
