+++ test/app.test.ts
@@ -1,25 +1,33 @@
+import Koa from 'koa';
 import parseLinkHeader from 'parse-link-header';
 import request from 'supertest';
-import app from '../src';
+import InMemoryArticles from '../src/adaptors/in-memory-articles';
+import createApp from '../src/app';
 
 const trim = (value: string): string => value.trim();
 const parseHeader = (header: string): Array<string> => header.split(',').map(trim);
 
+let app: Koa;
+
+beforeEach((): void => {
+  app = createApp({ articles: new InMemoryArticles() });
+});
+
 describe('the application', (): void => {
   it('should respond with 200 OK on the root', async (): Promise<void> => {
-    const response = await request(app).get('/');
+    const response = await request(app.callback()).get('/');
 
     expect(response.status).toEqual(200);
   });
 
   it('should respond with 404 Not Found on an unknown path', async (): Promise<void> => {
-    const response = await request(app).get('/does-not-exist');
+    const response = await request(app.callback()).get('/does-not-exist');
 
     expect(response.status).toEqual(404);
   });
 
   it('should support cross-origin requests', async (): Promise<void> => {
-    const response = await request(app).get('/').set('Origin', 'http://example.com');
+    const response = await request(app.callback()).get('/').set('Origin', 'http://example.com');
 
     expect(response.get('Access-Control-Allow-Origin')).toBe('http://example.com');
     expect(parseHeader(response.get('Vary'))).toContain('Origin');
@@ -27,7 +35,7 @@ describe('the application', (): void => {
   });
 
   it('should have an API documentation link', async (): Promise<void> => {
-    const response = await request(app).get('/');
+    const response = await request(app.callback()).get('/');
 
     expect(parseLinkHeader(response.get('Link'))).toHaveProperty(['http://www.w3.org/ns/hydra/core#apiDocumentation']);
   });
